requirements.txt
Django==4.1.2
djangorestframework==3.14.0
django-filter==22.1
psycopg2-binary==2.9.4
python-dotenv==0.21.0
pytest==7.1.3
pytest-django==4.5.2
model-bakery==1.8.0
pytest-cov==4.0.0